+++ component/module/file/file.go
@@ -11,6 +11,7 @@ import (
 	"github.com/grafana/agent/component/module"
 	"github.com/grafana/agent/service/cluster"
 	"github.com/grafana/agent/service/http"
+	otel_service "github.com/grafana/agent/service/otel"
 	"github.com/grafana/river/rivertypes"
 )
 
@@ -19,7 +20,7 @@ func init() {
 		Name:          "module.file",
 		Args:          Arguments{},
 		Exports:       module.Exports{},
-		NeedsServices: []string{http.ServiceName, cluster.ServiceName},
+		NeedsServices: []string{http.ServiceName, cluster.ServiceName, otel_service.ServiceName},
 
 		Build: func(opts component.Options, args component.Arguments) (component.Component, error) {
 			return New(opts, args.(Arguments))
@@ -91,7 +92,7 @@ func (c *Component) newManagedLocalComponent(o component.Options) (*file.Compone
 
 		if !c.inUpdate.Load() && c.isCreated.Load() {
 			// Any errors found here are reported via component health
-			_ = c.mod.LoadFlowContent(c.getArgs().Arguments, c.getContent().Value)
+			_ = c.mod.LoadFlowSource(c.getArgs().Arguments, c.getContent().Value)
 		}
 	}
 
@@ -138,7 +139,7 @@ func (c *Component) Update(args component.Arguments) error {
 
 	// Force a content load here and bubble up any error. This will catch problems
 	// on initial load.
-	return c.mod.LoadFlowContent(newArgs.Arguments, c.getContent().Value)
+	return c.mod.LoadFlowSource(newArgs.Arguments, c.getContent().Value)
 }
 
 // CurrentHealth implements component.HealthComponent.