+++ component/prometheus/exporter/mysql/mysql.go
@@ -1,11 +1,8 @@
 package mysql
 
 import (
-	"fmt"
-
 	"github.com/go-sql-driver/mysql"
 	"github.com/grafana/agent/component"
-	"github.com/grafana/agent/component/discovery"
 	"github.com/grafana/agent/component/prometheus/exporter"
 	"github.com/grafana/agent/pkg/integrations"
 	"github.com/grafana/agent/pkg/integrations/mysqld_exporter"
@@ -19,33 +16,13 @@ func init() {
 		Args:          Arguments{},
 		Exports:       exporter.Exports{},
 		NeedsServices: exporter.RequiredServices(),
-		Build:         exporter.NewWithTargetBuilder(createExporter, "mysql", customizeTarget),
+		Build:         exporter.New(createExporter, "mysql"),
 	})
 }
 
-func createExporter(opts component.Options, args component.Arguments) (integrations.Integration, error) {
+func createExporter(opts component.Options, args component.Arguments, defaultInstanceKey string) (integrations.Integration, string, error) {
 	a := args.(Arguments)
-	return a.Convert().NewIntegration(opts.Logger)
-}
-
-func customizeTarget(baseTarget discovery.Target, args component.Arguments) []discovery.Target {
-	a := args.(Arguments)
-	target := baseTarget
-
-	m, err := mysql.ParseDSN(string(a.DataSourceName))
-	if err != nil {
-		return []discovery.Target{target}
-	}
-
-	if m.Addr == "" {
-		m.Addr = "localhost:3306"
-	}
-	if m.Net == "" {
-		m.Net = "tcp"
-	}
-
-	target["instance"] = fmt.Sprintf("%s(%s)/%s", m.Net, m.Addr, m.DBName)
-	return []discovery.Target{target}
+	return integrations.NewIntegrationWithInstanceKey(opts.Logger, a.Convert(), defaultInstanceKey)
 }
 
 // DefaultArguments holds the default settings for the mysqld_exporter integration.