+++ pkg/integrations/install/install.go
@@ -31,6 +31,7 @@ import (
 	_ "github.com/grafana/agent/pkg/integrations/snowflake_exporter"     // register snowflake_exporter
 	_ "github.com/grafana/agent/pkg/integrations/squid_exporter"         // register squid_exporter
 	_ "github.com/grafana/agent/pkg/integrations/statsd_exporter"        // register statsd_exporter
+	_ "github.com/grafana/agent/pkg/integrations/vmware_exporter"        // register vmware_exporter
 	_ "github.com/grafana/agent/pkg/integrations/windows_exporter"       // register windows_exporter
 
 	//