+++ pkg/traces/config_test.go
@@ -615,6 +615,9 @@ exporters:
 processors:
   spanmetrics:
     metrics_exporter: prometheus
+    latency_histogram_buckets: {}
+    dimensions: {}
+
 extensions: {}
 service:
   pipelines:
@@ -981,6 +984,7 @@ processors:
     scrape_configs:
       - im_a_scrape_config
     operation_type: update
+    pod_associations: {}
 extensions: {}
 service:
   pipelines:
@@ -1022,130 +1026,6 @@ service:
     traces:
       exporters: ["otlp/0"]
       processors: ["service_graphs"]
-      receivers: ["push_receiver", "jaeger"]
-`,
-		},
-		{
-			name: "jaeger exporter",
-			cfg: `
-receivers:
-  jaeger:
-    protocols:
-      grpc:
-remote_write:
-  - insecure: true
-    format: jaeger
-    endpoint: example.com:12345
-`,
-			expectedConfig: `
-receivers:
-  push_receiver: {}
-  jaeger:
-    protocols:
-      grpc:
-exporters:
-  jaeger/0:
-    endpoint: example.com:12345
-    compression: gzip
-    tls:
-      insecure: true
-    retry_on_failure:
-      max_elapsed_time: 60s
-processors: {}
-extensions: {}
-service:
-  pipelines:
-    traces:
-      exporters: ["jaeger/0"]
-      processors: []
-      receivers: ["push_receiver", "jaeger"]
-`,
-		},
-		{
-			name: "jaeger exporter with basic auth",
-			cfg: `
-receivers:
-  jaeger:
-    protocols:
-      grpc:
-remote_write:
-  - insecure: true
-    format: jaeger
-    protocol: grpc
-    basic_auth:
-      username: test
-      password_file: ` + passwordFile.Name() + `
-    endpoint: example.com:12345
-`,
-			expectedConfig: `
-receivers:
-  push_receiver: {}
-  jaeger:
-    protocols:
-      grpc:
-exporters:
-  jaeger/0:
-    endpoint: example.com:12345
-    compression: gzip
-    tls:
-      insecure: true
-    headers:
-      authorization: Basic dGVzdDpwYXNzd29yZF9pbl9maWxl
-    retry_on_failure:
-      max_elapsed_time: 60s
-processors: {}
-extensions: {}
-service:
-  pipelines:
-    traces:
-      exporters: ["jaeger/0"]
-      processors: []
-      receivers: ["push_receiver", "jaeger"]
-`,
-		},
-		{
-			name: "two exporters different format",
-			cfg: `
-receivers:
-  jaeger:
-    protocols:
-      grpc:
-remote_write:
-  - insecure: true
-    format: jaeger
-    endpoint: example.com:12345
-  - insecure: true
-    format: otlp
-    endpoint: something.com:123
-`,
-			expectedConfig: `
-receivers:
-  push_receiver: {}
-  jaeger:
-    protocols:
-      grpc:
-exporters:
-  jaeger/0:
-    endpoint: example.com:12345
-    compression: gzip
-    tls:
-      insecure: true
-    retry_on_failure:
-      max_elapsed_time: 60s
-  otlp/1:
-    endpoint: something.com:123
-    compression: gzip
-    tls:
-      insecure: true
-    retry_on_failure:
-      max_elapsed_time: 60s
-processors: {}
-extensions: {}
-service:
-  pipelines:
-    traces:
-      exporters: ["jaeger/0", "otlp/1"]
-      processors: []
       receivers: ["push_receiver", "jaeger"]
 `,
 		},